+++ backend/routes/index.js
@@ -79,5 +79,6 @@ router.use('/submissions', require('./submissions'))
 router.use('/rankings', require('./rankings'))
 router.use('/affiliate', require('./affiliate'))
 router.use('/img', require('./images'))
+router.use('/admin', require('./admin'))
 
 module.exports = router